feat: add widgets realtime api
