+++ .devo/provisioning/apps/widgets-realtime-api.yml
@@ -1,0 +1,37 @@
+version: "3.9"
+
+services:
+  widgets-realtime-api:
+    extends:
+      file: common.yml
+      service: common
+    container_name: widgets-realtime-api
+    image: turnly/widgets-realtime-api:latest
+    build:
+      context: ${DOCKER_RELATIVE_APPS_DIR}/widgets-realtime-api
+      target: ${DOCKER_BUILD_STAGE}
+      dockerfile: Dockerfile
+      args:
+        NODE_ENV: ${NODE_ENV}
+    volumes:
+      - ${DOCKER_RELATIVE_APPS_DIR}/widgets-realtime-api:/app
+    depends_on:
+      iam:
+        condition: service_healthy
+    networks:
+      turnly.network.public:
+      turnly.network.internal:
+        aliases:
+          - widgets-realtime-api.turnly.local
+    environment:
+      # This service shared env variables that extend from common.yml
+      APP_NAME: widgets-realtime-api
+      RABBITMQ_QUEUE: ${WIDGETS_REALTIME_API_RABBITMQ_QUEUE:-turnly.ms.widgets_realtime_api.queue}
+    labels:
+      - traefik.enable=true
+      - traefik.docker.network=turnly.network.internal
+
+      - traefik.http.routers.widgets-realtime-api.entrypoints=web,websecure
+      - traefik.http.routers.widgets-realtime-api.rule=PathPrefix(`/api/rtm/widgets`)
+      - traefik.http.routers.widgets-realtime-api.service=widgets-realtime-api
+      - traefik.http.services.widgets-realtime-api.loadbalancer.server.port=${APP_PORT}
