--- apps/public-rtm-api/src/broadcasting/realtimeEvents.ts
@@ -1,26 +1,0 @@
-/**
- * Copyright (c) 2022, Turnly (https://turnly.app)
- * All rights reserved.
- *
- * Licensed under BSD 3-Clause License. See LICENSE for terms.
- */
-export enum RealtimeEvents {
-  /**
-   * Service events
-   */
-  SERVICE_TICKETS_AHEAD = 'service.tickets.ahead',
-  SERVICE_TICKETS_BEHIND = 'service.tickets.behind',
-
-  /**
-   * Ticket events
-   */
-  TICKET_BEFORE_YOURS_UPDATED = 'ticket.before-yours.updated',
-  TICKET_CALLED = 'ticket.called-to-desk',
-  TICKET_CANCELLED = 'ticket.cancelled',
-
-  /**
-   * Global events
-   */
-  CONNECTED = 'connected',
-  DISCONNECTED = 'disconnected',
-}
