+++ apps/widgets-realtime-api/.prettierignore
