+++ apps/widgets-realtime-api/src/Application.ts
@@ -4,9 +4,9 @@
  *
  * Licensed under BSD 3-Clause License. See LICENSE for terms.
  */
-import { Realtime } from '@turnly/realtime'
+import { BroadcastingEventsSubscriber, Realtime } from '@turnly/realtime'
 import { eventBus } from '@turnly/shared'
-import { BroadcastingEventsSubscriber } from 'broadcasting/BroadcastingEventsSubscriber'
+import { BROADCASTABLE_TO_CHANNELS, BroadcastableEvents } from 'broadcasting'
 import { queuingHandlers } from 'channels/queuing/handlers'
 import { AllowConnGuard } from 'channels/queuing/middlewares/AllowConnGuard'
 import { Channels, serverOptions } from 'shared/config'
@@ -46,7 +46,10 @@ export class Application {
    * @description Register broadcasting subscriber to event bus.
    */
   private setupBroadcastingEvents() {
-    const subscriber = new BroadcastingEventsSubscriber(this.realtime)
+    const subscriber = new BroadcastingEventsSubscriber<BroadcastableEvents>(
+      BROADCASTABLE_TO_CHANNELS,
+      this.realtime
+    )
 
     eventBus.subscribe([subscriber])
   }
