+++ apps/widgets-realtime-api/src/channels/queuing/handlers/RealtimeForCustomersHandler.ts
@@ -4,43 +4,33 @@
  *
  * Licensed under BSD 3-Clause License. See LICENSE for terms.
  */
-import { Guid } from '@turnly/common'
 import { AbstractRealtimeHandler, IRealtimeChannel } from '@turnly/realtime'
-import { Event, EventPayload, EventType } from '@turnly/shared'
+import { Event, EventType } from '@turnly/shared'
 import { BroadcastableEvents } from 'broadcasting/broadcastableEvents'
-import { RealtimeEvents } from 'broadcasting/realtimeEvents'
 
-interface Payload extends EventPayload {
-  serviceId: Guid
-  customerId: Guid
-  locationId: Guid
-}
+import { RealtimeEventsForQueuing } from '../events/RealtimeEventsForQueuing'
 
-export class RealtimeForCustomersHandler extends AbstractRealtimeHandler<
-  Event<Payload>
-> {
+export class RealtimeForCustomersHandler extends AbstractRealtimeHandler {
   public constructor() {
     super(RealtimeForCustomersHandler.getEvents())
   }
 
-  public handle(event: Event<Payload>, channel: IRealtimeChannel): void {
+  public handle(event: Event, channel: IRealtimeChannel): void {
     const name =
       event.getName() === BroadcastableEvents.TICKET_CREATED
-        ? RealtimeEvents.SERVICE_TICKETS_AHEAD
-        : RealtimeEvents.SERVICE_TICKETS_BEHIND
+        ? RealtimeEventsForQueuing.SERVICE_TICKETS_AHEAD
+        : RealtimeEventsForQueuing.SERVICE_TICKETS_BEHIND
 
-    const payload = {
-      locationId: event.payload.locationId,
-      serviceId: event.payload.serviceId,
-      organizationId: event.payload.organizationId,
-    }
-
-    channel.to(payload.locationId).emit(
+    channel.to(event.payload.locationId).emit(
       name,
       Event.fromObject({
         type: EventType.INFO,
         name,
-        payload,
+        payload: {
+          locationId: event.payload.locationId,
+          serviceId: event.payload.serviceId,
+          organizationId: event.payload.organizationId,
+        },
       })
     )
   }
