+++ apps/widgets-realtime-api/src/channels/queuing/handlers/RealtimeTicketCalledToDeskHandler.ts
@@ -8,7 +8,8 @@ import { Guid } from '@turnly/common'
 import { AbstractRealtimeHandler, IRealtimeChannel } from '@turnly/realtime'
 import { Event, EventPayload, EventType } from '@turnly/shared'
 import { BroadcastableEvents } from 'broadcasting/broadcastableEvents'
-import { RealtimeEvents } from 'broadcasting/realtimeEvents'
+
+import { RealtimeEventsForQueuing } from '../events/RealtimeEventsForQueuing'
 
 interface Payload extends EventPayload {
   id: Guid
@@ -16,11 +17,11 @@ interface Payload extends EventPayload {
   status: string
 }
 
-export class RealtimeTicketCancelledHandler extends AbstractRealtimeHandler<
+export class RealtimeTicketCalledToDeskHandler extends AbstractRealtimeHandler<
   Event<Payload>
 > {
   public constructor() {
-    super(RealtimeTicketCancelledHandler.getEvents())
+    super(RealtimeTicketCalledToDeskHandler.getEvents())
   }
 
   public handle(
@@ -30,10 +31,10 @@ export class RealtimeTicketCancelledHandler extends AbstractRealtimeHandler<
     channel: IRealtimeChannel
   ): void {
     channel.to(customerId).emit(
-      RealtimeEvents.TICKET_CANCELLED,
+      RealtimeEventsForQueuing.TICKET_CALLED,
       Event.fromObject({
         type: EventType.INFO,
-        name: RealtimeEvents.TICKET_CANCELLED,
+        name: RealtimeEventsForQueuing.TICKET_CALLED,
         payload: {
           ticketId,
           status,
@@ -44,11 +45,6 @@ export class RealtimeTicketCancelledHandler extends AbstractRealtimeHandler<
   }
 
   public static getEvents(): BroadcastableEvents[] {
-    return [
-      BroadcastableEvents.TICKET_COMPLETED,
-      BroadcastableEvents.TICKET_CANCELLED,
-      BroadcastableEvents.TICKET_RETURNED,
-      BroadcastableEvents.TICKET_DISCARDED,
-    ]
+    return [BroadcastableEvents.TICKET_CALLED]
   }
 }
