+++ apps/widgets-realtime-api/src/channels/queuing/middlewares/AllowConnGuard.ts
