+++ apps/widgets-realtime-api/tsconfig.json
